KIEKER-1885 added more debugging info to identify race condition.
