+++ kieker-analysis/src/kieker/analysis/ExampleConfiguration.java
@@ -30,16 +30,6 @@ import kieker.analysis.graph.dependency.dot.DotExportConfigurationFactory;
 import kieker.analysis.graph.dependency.vertextypes.IVertexTypeMapper;
 import kieker.analysis.graph.export.dot.DotExportConfiguration;
 import kieker.analysis.graph.export.dot.DotFileWriterStage;
-import kieker.analysis.model.CallEvent2OperationCallStage;
-import kieker.analysis.model.ExecutionModelAssembler;
-import kieker.analysis.model.ExecutionModelAssemblerStage;
-import kieker.analysis.model.ModelObjectFromOperationCallAccessors;
-import kieker.analysis.model.ModelRepository;
-import kieker.analysis.model.OperationAndCallGeneratorStage;
-import kieker.analysis.model.OperationCallExtractorStage;
-import kieker.analysis.model.StaticModelsAssemblerStage;
-import kieker.analysis.model.data.CallEvent;
-import kieker.analysis.model.data.OperationEvent;
 import kieker.analysis.signature.NameBuilder;
 import kieker.analysis.signature.SignatureExtractor;
 import kieker.analysis.source.file.DirectoryReaderStage;
@@ -47,6 +37,16 @@ import kieker.analysis.source.file.DirectoryScannerStage;
 import kieker.analysis.stage.flow.FlowTraceEventMatcher;
 import kieker.analysis.stage.general.CallEventMatcher;
 import kieker.analysis.stage.general.ControlledEventReleaseStage;
+import kieker.analysis.stage.model.CallEvent2OperationCallStage;
+import kieker.analysis.stage.model.ExecutionModelAssembler;
+import kieker.analysis.stage.model.ExecutionModelAssemblerStage;
+import kieker.analysis.stage.model.ModelObjectFromOperationCallAccessors;
+import kieker.analysis.stage.model.ModelRepository;
+import kieker.analysis.stage.model.OperationAndCallGeneratorStage;
+import kieker.analysis.stage.model.OperationCallExtractorStage;
+import kieker.analysis.stage.model.StaticModelsAssemblerStage;
+import kieker.analysis.stage.model.data.CallEvent;
+import kieker.analysis.stage.model.data.OperationEvent;
 import kieker.analysis.statistics.CallStatisticsStage;
 import kieker.analysis.statistics.FullResponseTimeStatisticsStage;
 import kieker.analysis.trace.graph.TraceToGraphTransformerStage;
