+++ kieker-analysis/src/kieker/analysis/debug/hotspotdetection/HotspotDetectionConfiguration.java
@@ -19,17 +19,17 @@ package kieker.analysis.debug.hotspotdetection;
 import java.io.File;
 import java.time.temporal.ChronoUnit;
 
-import kieker.analysis.model.AssemblyModelAssemblerStage;
-import kieker.analysis.model.DeploymentModelAssemblerStage;
-import kieker.analysis.model.OperationAndCallGeneratorStage;
-import kieker.analysis.model.TypeModelAssemblerStage;
-import kieker.analysis.model.data.OperationEvent;
 import kieker.analysis.signature.JavaComponentSignatureExtractor;
 import kieker.analysis.signature.JavaOperationSignatureExtractor;
 import kieker.analysis.source.file.DirectoryReaderStage;
 import kieker.analysis.source.file.DirectoryScannerStage;
 import kieker.analysis.stage.flow.FlowTraceEventMatcher;
 import kieker.analysis.stage.general.ControlledEventReleaseStage;
+import kieker.analysis.stage.model.AssemblyModelAssemblerStage;
+import kieker.analysis.stage.model.DeploymentModelAssemblerStage;
+import kieker.analysis.stage.model.OperationAndCallGeneratorStage;
+import kieker.analysis.stage.model.TypeModelAssemblerStage;
+import kieker.analysis.stage.model.data.OperationEvent;
 import kieker.analysis.trace.reconstruction.FlowRecordTraceReconstructionStage;
 import kieker.common.record.IMonitoringRecord;
 import kieker.common.record.flow.IFlowRecord;
