+++ kieker-analysis/src/kieker/analysis/graph/dependency/IDependencyGraphBuilderFactory.java
@@ -16,7 +16,7 @@
 
 package kieker.analysis.graph.dependency;
 
-import kieker.analysis.model.ModelRepository;
+import kieker.analysis.stage.model.ModelRepository;
 
 /**
  * @author Sören Henning
