+++ kieker-analysis/src/kieker/analysis/stage/model/DeploymentModelPrinter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  ***************************************************************************/
 
-package kieker.analysis.model;
+package kieker.analysis.stage.model;
 
 import java.io.PrintStream;
