+++ kieker-analysis/src/kieker/analysis/stage/model/ExecutionModelAssemblerStage.java
@@ -1,5 +1,5 @@
 /***************************************************************************
- * Copyright 2021 Kieker Project (http://kieker-monitoring.net)
+ * Copyright 2020 Kieker Project (http://kieker-monitoring.net)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package kieker.analysis.model;
 
-import kieker.model.analysismodel.trace.OperationCall;
+package kieker.analysis.stage.model;
+
+import kieker.model.analysismodel.deployment.DeployedOperation;
+import kieker.model.analysismodel.execution.Tuple;
+
+import teetime.stage.basic.AbstractFilter;
 
 /**
  * @author Reiner Jung
- *
  * @since 1.15
- *
  */
-public interface ITraceBasedExecutionModelAssembler {
+public class ExecutionModelAssemblerStage extends AbstractFilter<Tuple<DeployedOperation, DeployedOperation>> {
 
-	void addOperationCall(final OperationCall operationCall);
+	private final IExecutionModelAssembler assembler;
+
+	public ExecutionModelAssemblerStage(final IExecutionModelAssembler assembler) {
+		this.assembler = assembler;
+	}
+
+	@Override
+	protected void execute(final Tuple<DeployedOperation, DeployedOperation> operationCall) {
+		this.assembler.addOperationCall(operationCall);
+		this.outputPort.send(operationCall);
+	}
 
 }
