+++ kieker-analysis/src/kieker/analysis/stage/model/MetaModelHandler.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  ***************************************************************************/
 
-package kieker.analysis.model;
+package kieker.analysis.stage.model;
 
 import java.io.File;
 import java.io.IOException;
