+++ kieker-analysis/src/kieker/analysis/stage/model/ModelObjectFromOperationCallAccessors.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  ***************************************************************************/
 
-package kieker.analysis.model;
+package kieker.analysis.stage.model;
 
 import java.util.function.Function;
