+++ kieker-analysis/test/kieker/analysis/ExampleConfigurationTest.java
@@ -31,7 +31,7 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import kieker.analysis.model.DeploymentModelPrinter;
+import kieker.analysis.stage.model.DeploymentModelPrinter;
 import kieker.model.analysismodel.deployment.DeployedOperation;
 import kieker.model.analysismodel.deployment.DeploymentModel;
 import kieker.model.analysismodel.execution.AggregatedInvocation;
