+++ kieker-analysis/test/kieker/analysis/dependencygraphs/DeploymentLevelComponentDependencyGraphBuilderFactoryTest.java
@@ -24,7 +24,7 @@ import org.junit.Test;
 import kieker.analysis.graph.dependency.DeploymentLevelComponentDependencyGraphBuilder;
 import kieker.analysis.graph.dependency.DeploymentLevelComponentDependencyGraphBuilderFactory;
 import kieker.analysis.graph.dependency.IDependencyGraphBuilder;
-import kieker.analysis.model.ModelRepository;
+import kieker.analysis.stage.model.ModelRepository;
 import kieker.model.analysismodel.execution.ExecutionFactory;
 import kieker.model.analysismodel.execution.ExecutionModel;
 import kieker.model.analysismodel.statistics.StatisticsFactory;
